requirements.txt
pillow
Levenshtein
langchain
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-gpu
langchain-community
langchain-huggingface
fitz
pymupdf